+++ demo/lit-code-example.js
@@ -2,33 +2,40 @@ export default `import { LitElement, html } from 'lit-element';
 import '@anypoint-web-components/anypoint-signin/anypoint-signin.js';
 
 class SampleElement extends LitElement {
-  constructor() {
-    super();
-    ...
-    this.exchangeCode = this.exchangeCode.bind(this);
+  /**
+   * @param {AnypointCodeExchangeEvent} e
+   */
+  codeEventHandler(e) {
+    const { code } = e;
+    e.detail.result = this.exchangeCode(code);
   }
 
-  connectedCallback() {
-    super.connectedCallback();
-    window.addEventListener('oauth2-code-response', this.exchangeCode);
-  }
-
-  disconnectedCallback() {
-    super.disconnectedCallback();
-    window.removeEventListener('oauth2-code-response', this.exchangeCode);
-  }
-
-  async exchangeCode(e) {
-    const { code } = e.detail;
+  /**
+   * @param {string} code
+   * @returns {Promise<TokenInfo>}
+   */
+  async exchangeCode(code) {
     const init = {
       method: 'POST',
-      body: code
+      body: code,
     };
-    const tokenExchangeUrl = 'YOUR SERVER URL';
-    const response = await fetch(tokenExchangeUrl, init);
-    this.token = await response.json();
-    const button = this.shadowRoot.querySelector('anypoint-signin');
-    button.signedIn = !!this.token;
+    const response = await fetch('YOUR SERVER URL', init);
+    const info = await response.json();
+    return {
+      accessToken: info.accessToken,
+      expiresAt: Date.now() + info.expiresIn,
+      expiresIn: info.expiresIn,
+      expiresAssumed: false,
+      state: '0', // this is required by the types definition but can be anything. State is checked before this function is called
+    };
+  }
+
+  render() {
+    return html\`
+      <anypoint-signin
+        @anypointcodeexchange="\${this.codeEventHandler}"
+      ></anypoint-signin>
+    \`;
   }
 }
 customElements.define('sample-element', SampleElement);