+++ src/internals.d.ts
@@ -1,0 +1,6 @@
+export declare const clientIdValue: unique symbol;
+export declare const redirectUriValue: unique symbol;
+export declare const authTypeValue: unique symbol;
+export declare const signedInValue: unique symbol;
+export declare const accessTokenValue: unique symbol;
+export declare const scopesValue: unique symbol;