+++ src/internals.js
@@ -1,0 +1,6 @@
+export const clientIdValue = Symbol('clientIdValue');
+export const redirectUriValue = Symbol('redirectUriValue');
+export const authTypeValue = Symbol('authTypeValue');
+export const signedInValue = Symbol('signedInValue');
+export const accessTokenValue = Symbol('accessTokenValue');
+export const scopesValue = Symbol('scopesValue');